Philip添加提交订单的Controller，将获取到的信息并向数据库插入Order对象

+++ src/main/java/com/cake/pojo/Order.java
@@ -12,9 +12,9 @@ public class Order {
 
     private Integer amount;//商品总数
 
-    private Byte status;//订单状态(1未付款/2已付款/3已发货/4已完成)
+    private Integer status;//订单状态(1未付款/2已付款/3已发货/4已完成)
 
-    private Byte paytype;//支付方式 (1微信/2支付宝/3货到付款)
+    private Integer paytype;//支付方式 (1微信/2支付宝/3货到付款)
 
     private String name;//收货人
 
@@ -26,7 +26,7 @@ public class Order {
 
     private Integer userId;//下单用户
 
-    public Order(Integer id, Integer total, Integer amount, Byte status, Byte paytype, String name, String phone, String address, Date systime, Integer userId) {
+    public Order(Integer id, Integer total, Integer amount, Integer status, Integer paytype, String name, String phone, String address, Date systime, Integer userId) {
         this.id = id;
         this.total = total;
         this.amount = amount;
@@ -67,19 +67,19 @@ public class Order {
         this.amount = amount;
     }
 
-    public Byte getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(Byte status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
 
-    public Byte getPaytype() {
+    public Integer getPaytype() {
         return paytype;
     }
 
-    public void setPaytype(Byte paytype) {
+    public void setPaytype(Integer paytype) {
         this.paytype = paytype;
     }
